api/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
python-multipart
requests
python-telegram-bot
